WL: fully type wlrq module
