+++ libqtile/backend/wayland/wlrq.py
@@ -28,26 +28,20 @@ import cairocffi
 from pywayland.server import Listener
 from wlroots.wlr_types import Texture
 from wlroots.wlr_types.keyboard import KeyboardModifier
-from wlroots.wlr_types.pointer_constraints_v1 import (
-    PointerConstraintV1,
-    PointerConstraintV1StateField,
-)
-from wlroots.wlr_types.xdg_shell import XdgSurface
 
-from libqtile.backend.base import Internal
 from libqtile.log_utils import logger
 from libqtile.utils import QtileError
 
 if TYPE_CHECKING:
-    from typing import Callable
+    from typing import Any, Callable
 
     from pywayland.server import Signal
     from wlroots import xwayland
-    from wlroots.wlr_types import Box, data_device_manager
+    from wlroots.wlr_types import data_device_manager
 
     from libqtile.backend.wayland.core import Core
     from libqtile.backend.wayland.output import Output
-    from libqtile.backend.wayland.window import WindowType
+    from libqtile.config import Screen
 
 
 class WlrQError(QtileError):
@@ -112,10 +106,10 @@ def translate_masks(modifiers: list[str]) -> int:
 
 
 class Painter:
-    def __init__(self, core):
+    def __init__(self, core: Core):
         self.core = core
 
-    def paint(self, screen, image_path, mode=None):
+    def paint(self, screen: Screen, image_path: str, mode: str | None = None) -> None:
         try:
             with open(image_path, "rb") as f:
                 image, _ = cairocffi.pixbuf.decode_to_image_surface(f.read())
@@ -155,7 +149,10 @@ class Painter:
                 screen.height,
                 cairocffi.cairo.cairo_image_surface_get_data(surface._pointer),
             )
-            outputs = [output for output in self.core.outputs if output.wlr_output.enabled]
+            # mypy struggles to understand this. See: https://github.com/python/mypy/issues/11513
+            outputs = [
+                output for output in self.core.outputs if output.wlr_output.enabled  # type: ignore
+            ]
             outputs[screen.index].wallpaper = texture
 
 
@@ -167,92 +164,16 @@ class HasListeners:
     This guarantees that all listeners that set up and then removed in reverse order.
     """
 
-    def add_listener(self, event: Signal, callback: Callable):
+    def add_listener(self, event: Signal, callback: Callable) -> None:
         if not hasattr(self, "_listeners"):
             self._listeners = []
         listener = Listener(callback)
         event.add(listener)
         self._listeners.append(listener)
 
-    def finalize_listeners(self):
+    def finalize_listeners(self) -> None:
         for listener in reversed(self._listeners):
             listener.remove()
-
-
-class PointerConstraint(HasListeners):
-    """
-    A small object to listen to signals on `struct wlr_pointer_constraint_v1` instances.
-    """
-
-    rect: Box
-
-    def __init__(self, core: Core, wlr_constraint: PointerConstraintV1):
-        self.core = core
-        self.wlr_constraint = wlr_constraint
-        self.window: WindowType | None = None
-        self._warp_target = (0, 0)
-        self._needs_warp = False
-
-        self.add_listener(wlr_constraint.set_region_event, self._on_set_region)
-        self.add_listener(wlr_constraint.destroy_event, self._on_destroy)
-
-        self._get_window()
-
-    def _get_window(self):
-        for win in self.core.qtile.windows_map.values():
-            if not isinstance(win, Internal) and isinstance(win.surface, XdgSurface):
-                if win.surface.surface == self.wlr_constraint.surface:
-                    break
-        else:
-            self.finalize()
-
-        self.window = win
-
-    def finalize(self):
-        if self.core.active_pointer_constraint is self:
-            self.disable()
-        self.finalize_listeners()
-        self.core.pointer_constraints.remove(self)
-
-    def _on_set_region(self, _listener, _data):
-        logger.debug("Signal: wlr_pointer_constraint_v1 set_region")
-        self._get_region()
-
-    def _on_destroy(self, _listener, wlr_constraint: PointerConstraintV1):
-        logger.debug("Signal: wlr_pointer_constraint_v1 destroy")
-        self.finalize()
-
-    def _on_commit(self, _listener, _data):
-        if self._needs_warp:
-            # Warp in case the pointer is not inside the rect
-            if not self.rect.contains_point(self.cursor.x, self.cursor.y):
-                self.core.warp_pointer(*self._warp_target)
-            self._needs_warp = False
-
-    def _get_region(self):
-        rect = self.wlr_constraint.region.rectangles_as_boxes()[0]
-        rect.x += self.window.x + self.window.borderwidth
-        rect.y += self.window.y + self.window.borderwidth
-        self._warp_target = (rect.x + rect.width / 2, rect.y + rect.height / 2)
-        self.rect = rect
-        self._needs_warp = True
-
-    def enable(self):
-        logger.debug("Enabling pointer constraints.")
-        self.core.active_pointer_constraint = self
-        self._get_region()
-        self.add_listener(self.wlr_constraint.surface.commit_event, self._on_commit)
-        self.wlr_constraint.send_activated()
-
-    def disable(self):
-        logger.debug("Disabling pointer constraints.")
-
-        if self.wlr_constraint.current.committed & PointerConstraintV1StateField.CURSOR_HINT:
-            x, y = self.wlr_constraint.current.cursor_hint
-            self.core.warp_pointer(x + self.window.x, y + self.window.y)
-
-        self.core.active_pointer_constraint = None
-        self.wlr_constraint.send_deactivated()
 
 
 class Dnd(HasListeners):
@@ -278,24 +199,24 @@ class Dnd(HasListeners):
         self.finalize_listeners()
         self.core.live_dnd = None
 
-    def _on_destroy(self, _listener, _event) -> None:
+    def _on_destroy(self, _listener: Listener, _event: Any) -> None:
         logger.debug("Signal: wlr_drag destroy")
         self.finalize()
 
-    def _on_icon_map(self, _listener, _event) -> None:
+    def _on_icon_map(self, _listener: Listener, _event: Any) -> None:
         logger.debug("Signal: wlr_drag_icon map")
         for output in self._outputs:
             output.damage()
 
-    def _on_icon_unmap(self, _listener, _event) -> None:
+    def _on_icon_unmap(self, _listener: Listener, _event: Any) -> None:
         logger.debug("Signal: wlr_drag_icon unmap")
         for output in self._outputs:
             output.damage()
 
-    def _on_icon_destroy(self, _listener, _event) -> None:
+    def _on_icon_destroy(self, _listener: Listener, _event: Any) -> None:
         logger.debug("Signal: wlr_drag_icon destroy")
 
-    def _on_icon_commit(self, _listener, _event) -> None:
+    def _on_icon_commit(self, _listener: Listener, _event: Any) -> None:
         self.width = self.wlr_drag.icon.surface.current.width
         self.height = self.wlr_drag.icon.surface.current.height
         self.position(self.core.cursor.x, self.core.cursor.y)
